LPS-76246 Interface is not necessary. Convert it to a component without an interface
